fixes issue #16
add regression test for moved classes.
